+++ app/src/main/java/de/passwordvault/frontend/main/entries/ListAdapter.java
@@ -1,38 +1,29 @@
 package de.passwordvault.frontend.main.entries;
 
 import android.content.Context;
-import android.content.Intent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.animation.Animation;
 import android.view.animation.AnimationUtils;
-import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
-
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.recyclerview.widget.RecyclerView;
-
 import java.util.ArrayList;
-import java.util.Objects;
-
 import de.passwordvault.R;
-import de.passwordvault.backend.Singleton;
-import de.passwordvault.backend.entry.AbbreviatedEntry;
-import de.passwordvault.frontend.entry.EntryActivity;
+import de.passwordvault.backend.entry.Entry;
 
 
 /**
- * Class models a custom {@linkplain ArrayAdapter} for {@linkplain AbbreviatedEntry} instances which
+ * Class models a custom {@linkplain ArrayAdapter} for {@linkplain Entry} instances which
  * are displayed through a {@linkplain android.widget.ListView} within the {@linkplain EntriesFragment}.
  *
  * @author  Christian-2003
- * @version 1.0.0
+ * @version 2.0.0
  */
 // implements AdapterView.OnItemClickListener
-public class ListAdapter extends ArrayAdapter<AbbreviatedEntry> {
+public class ListAdapter extends ArrayAdapter<Entry> {
 
     /**
      * Attribute stores the context for the adapter.
@@ -51,7 +42,7 @@ public class ListAdapter extends ArrayAdapter<AbbreviatedEntry> {
      * @param entries   AbbreviatedEntries to be displayed with the ListView.
      * @param context   Context for the adapter.
      */
-    public ListAdapter(ArrayList<AbbreviatedEntry> entries, Context context) {
+    public ListAdapter(ArrayList<Entry> entries, Context context) {
         super(context, R.layout.entries_list_item, entries);
         this.context = context;
         lastPosition = -1;
@@ -69,7 +60,7 @@ public class ListAdapter extends ArrayAdapter<AbbreviatedEntry> {
     @NonNull
     @Override
     public View getView(int position, @Nullable View convertView, @NonNull ViewGroup parent) {
-        AbbreviatedEntry entry = getItem(position);
+        Entry entry = getItem(position);
         LayoutInflater inflater = LayoutInflater.from(getContext());
         convertView = inflater.inflate(R.layout.entries_list_item, parent, false);
         ((TextView)convertView.findViewById(R.id.entries_list_item_name)).setText(entry.getName());